requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
pymongo>=3.12.0
python-multipart>=0.0.5
minio>=7.1.0
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytesseract
opencv-python-headless
numpy
pillow 